CrazyFun: Remaining Namespace work (Aside from the .rl items) (#7736)
Namespace all of RubyMappings classes, tasks (and make inheritance chain correct), and RakeMappings (in the same vein as RubyMappings)
