+++ rake_tasks/crazy_fun/mappings/ruby_mappings.rb
@@ -8,20 +8,24 @@ require_relative 'ruby_mappings/ruby_library'
 require_relative 'ruby_mappings/ruby_linter'
 require_relative 'ruby_mappings/ruby_test'
 
-class RubyMappings
-  def add_all(fun)
-    fun.add_mapping "ruby_library", RubyLibrary.new
+module CrazyFun
+  module Mappings
+    class RubyMappings
+      def add_all(fun)
+        fun.add_mapping "ruby_library", RubyLibrary.new
 
-    fun.add_mapping "ruby_test", CheckTestArgs.new
-    fun.add_mapping "ruby_test", AddTestDefaults.new
-    fun.add_mapping "ruby_test", ExpandSourceFiles.new
-    fun.add_mapping "ruby_test", RubyTest.new
-    fun.add_mapping "ruby_test", AddTestDependencies.new
+        fun.add_mapping "ruby_test", CheckTestArgs.new
+        fun.add_mapping "ruby_test", AddTestDefaults.new
+        fun.add_mapping "ruby_test", ExpandSourceFiles.new
+        fun.add_mapping "ruby_test", RubyTest.new
+        fun.add_mapping "ruby_test", AddTestDependencies.new
 
-    fun.add_mapping "ruby_lint", ExpandSourceFiles.new
-    fun.add_mapping "ruby_lint", RubyLinter.new
+        fun.add_mapping "ruby_lint", ExpandSourceFiles.new
+        fun.add_mapping "ruby_lint", RubyLinter.new
 
-    fun.add_mapping "rubydocs", RubyDocs.new
-    fun.add_mapping "rubygem", RubyGem.new
+        fun.add_mapping "rubydocs", RubyDocs.new
+        fun.add_mapping "rubygem", RubyGem.new
+      end
+    end
   end
 end
