+++ rake_tasks/crazy_fun/mappings/ruby_mappings/add_test_dependencies.rb
@@ -1,10 +1,14 @@
-class RubyMappings
-  class AddTestDependencies < Tasks
-    def handle(_fun, dir, args)
-      task = Rake::Task[task_name(dir, "#{args[:name]}-test")]
+module CrazyFun
+  module Mappings
+    class RubyMappings
+      class AddTestDependencies < CrazyFun::Mappings::Tasks
+        def handle(_fun, dir, args)
+          task = Rake::Task[task_name(dir, "#{args[:name]}-test")]
 
-      if args.has_key?(:deps)
-        add_dependencies task, dir, args[:deps]
+          if args.has_key?(:deps)
+            add_dependencies task, dir, args[:deps]
+          end
+        end
       end
     end
   end
