+++ rake_tasks/crazy_fun/mappings/ruby_mappings/expand_source_files.rb
@@ -1,7 +1,11 @@
-class RubyMappings
-  class ExpandSourceFiles
-    def handle(_fun, dir, args)
-      args[:srcs] = args[:srcs].map { |str| Dir[File.join(dir, str)] }.flatten
+module CrazyFun
+  module Mappings
+    class RubyMappings
+      class ExpandSourceFiles
+        def handle(_fun, dir, args)
+          args[:srcs] = args[:srcs].map { |str| Dir[File.join(dir, str)] }.flatten
+        end
+      end
     end
   end
 end
